requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
dj-database-url
python-dotenv
cloudinary
django-cloudinary-storage
